ci: Don't try to overwrite immutable operator OCIs (#3320)
Check whether the destination operator OCI exists before trying to build
and push it. The ECR repository is typically immutable so the push will
fail if the image exists. This happens when another part of the tag
build fails after the operator OCI has been built, and we re-run the
`build-operator-oci` workflow. This prevents forward progress with a
release.

This problem has existed for a while but has become more apparent now
the Drone pipeline that calls this workflow no longer calls the workflow
to build the legacy OCIs. This means the Drone pipeline no longer
depends on previous build pipelines so the operator OCI images are built
right from the start. As it completes quickly, any failure in another
pipeline means we try to rebuild the operator OCI, which then fails.

Source-commit: gravitational/teleport.e@a11fa430691d42cdae4d8e4f6e1e280b78542aa7
Source-pull-requests: gravitational/teleport.e#3320