AWS OIDC: remove dependency on servicecfg
Currently the ExternalAuditStorage methods require the `servicecfg` package
to re-use the configuration struct used in `teleport integration configure`.

We are refactoring the AWS OIDC actions to use a gRPC Service.
Importing the `sevicecfg` package makes this refactoring a bit difficult
because of a cycling dependency: `awsoidc` requires `servicecfg` which
requires `auth` which will require the `awsoidc service` which requires
the `awsoidc` package.

This PR removes the usage of the `servicecfg` package in `awsoidc` package.