[v14] AWS OIDC: remove config dependency in DeployService (#37816)
* AWS OIDC: remove config dependency

Currently the DeployService and DeployDatabaseService require the
`config` package to create a `teleport.yaml` configuration that is used
to start the `teleport` process on the ECS containers.

We are refactoring the AWS OIDC actions to use a gRPC Service.
Importing the `config` package makes this refactoring a bit difficult
because of a cyclic dependency.

This PR removes the usage of the `config` package in `awsoidc` package.

* rename field name