ci: Add GitHub action for parsing semver (#3290)
* ci: Add GitHub action for parsing semver

Add a GitHub composite action for parsing semver strings for:
* validating conformance with semver.org,
* extracting the components of a version string,
* making a label list with X.Y.Z, X.Y and X labels for non-pre-release
  versions, to be used to label container images, etc,
* providing a version override for when a git ref may not be a tag,
  allowing a version to be provided for test builds when the ref is not
  a release tag.

Validation allows a leading `v`, but that is stripped from the output
values. This allows a tag ref to be passed in as-is and still pass
validation.

* ci: Add tests for parse-version action

Add a new composite action to run a single unit test of the
parse-version action, and a top-level workflow to run multiple unit
tests, to be triggered by a `workflow_dispatch` event.

Source-commit: gravitational/teleport.e@7c9ed8f5be709e65066c8e58bde4d63b8156e4da
Source-pull-requests: gravitational/teleport.e#3290