ci: Add environment input to operator OCI build (#3372)
Add an input parameter for environment to the build of the operator OCI
and use the vars from the environment instead of taking inputs with
defaults. This will allows us to build in the staging environment which
has recently been set up with values for staging.

Source-commit: gravitational/teleport.e@22e6764de05c4a2454aa24c6c3c469e06229f3e8
Source-pull-requests: gravitational/teleport.e#3372