ci: Add workflow for publishing (promoting) Helm charts (#3288)
* ci: Add workflow for publishing (promoting) Helm charts

Add a workflow to supersede the part of the Drone pipeline
`promote-build` that updates the helm repository with the charts in the
version being published. This is currently intended to be called from
Drone, and will subsequently be called by the top-level `tag-publish`
workflow.

* ci: Publish teleport operator helm chart

Publish the teleport operator helm chart as part of publishing the other
helm charts. This is a separate commit as v14 and earlier do not publish
the operator helm chart, and a separate commit will make backporting
easier.

This mirrors the initial addition done to Drone in
gravitational/teleport@018668b2a620.

Source-commit: gravitational/teleport.e@0ca47603962bd916419600b41766bde426d64bc5
Source-pull-requests: gravitational/teleport.e#3288