attempt better admin permission handling
With this commit, you can start VNet as a regular non-root user, and it
will fork itself as root via an applescript wrapper (that prompts for
the user's password) to do the necessary setup.
This means you don't need to run tsh vnet as root, and it won't write
root-owned files into your ~/.tsh directory.

Currently this is passing file descriptors through unix sockets and it
feels very hacky, I don't love it, might be better to start as root and
fork as non-root later.