Fix panic when `TLSServer.Close()` is invoked before `TLSServer.Serve()`
This PR prevents a panic when the `TLSServer` is closed before the
`Serve` function is called. This happens in special situations during
tests when something else failed causing the test to fail which triggers
the cleanup function.
If the cleanup function executes before `Serve` goroutine is executed,
the listener is nil which triggers the panic.

Part of #37884

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>