Update `cert-manager` instructions in Teleport cluster guides (#37891)
This commit updates the `cert-manager` installation instructions to configure the `cert-manager` leader election to use `cert-manager` namespace instead of `kube-system`.

This is required because in GCP Autopilot clusters lock down access to `kube-system` and helm installation can't be successful.

By default, `cert-manager` uses `kube-system` as leader election to prevent multiple parallel installations of `cert-manager` in the same cluster, but if it's already running in the target cluster, the installation will fail with a non clear error. Given the above, it's preferred to use a separate namespace for leader election to prevent races.

Fixes #25312

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>