build: Update buildbox definitions for spacelift runner
Add a var to `build.assets/images.mk` that defines the centos7 buildbox
without the host architecture suffix, and print that in the
`show-buildbox-centos7-target` used by the spacelift runner CI job.

The spacelift runner needs to know the buildbox to use to build `tbot`,
but since it does a `docker buildx` multiarch build, it needs to be able
to select different architecture buildboxes itself using `$TARGETARCH`
and not just use the one for the host architecture.