Attempts to stabilize lib/srv/regular tests
All tests that use the common fixture suffer from periodic failures
due to `timeout waiting for announce to be sent`. This stems from
the fixture forcing a heartbeat for the node being created in the
test timing out. Instead of using the heartbeat interface to fire
an early heartbeat, the server is upserted directly into the test
auth instance.