Prevent leaking gRPC clients when connecting to the cluster fails (#38003)
There are a few scenarios during agent registration that may result
in an auth client being created but never closed. One particular
scenario that can hit this path is a v12 agent trying to join a
>=v14 control plane. The first thing agents do after validating
that the client was created succesfully is retrieve the Host CA
for the cluster. If that fails then the error is returned but
the client is never closed and the agent will try connecting again
in the future.

This behavior is particularly troublesome because in v12 the API
responsible for that was the auth HTTP API, in V13 it was
deprecated in favor of a gRPC API and in v14 the HTTP API was
removed entirely. So any v12 agents trying to connect to a >=v14
control plane will continuously fail getting the Host CA because the
HTTP request they issue will result in a 404.