Remove rdpclient build from Go tests (#38098)
In CI we run parallel make, and as a result the linker sometimes
tries to link librdp_client.a before the Rust build has completed.

We could fix this, but our unit tests for Rust and Go are separate
and there are no Go tests that require the Rust library to be present,
so for now we remove the Rust build as a prerequisite.

This should fix the random test failures and also improve the speed
of all test runs.