ci: Update checkout-teleport action to use actions/github-script@v7 (#3402)
Update the version of `actions/github-script` to v7 from v6.

Using v6 results in the following warning on a GitHub Actions workflow
run overview:

    Node.js 16 actions are deprecated. Please update the following
    actions to use Node.js 20: actions/github-script@v6. For more
    information see:
    https://github.blog/changelog/2023-09-22-github-actions-transitioning-from-node-16-to-node-20/.

This makes the action unusable by workflows that use centos7 as a
container. We do not use that container for workflow jobs/steps in this
teleport.e repository, and this action is only used in teleport.e.

Source-commit: gravitational/teleport.e@78aca5e4293bbe00cb2c8821a1f120e2759b997e
Source-pull-requests: gravitational/teleport.e#3402