Set noEmit in tsconfig.json (#37999)
We use Vite to compile TS files, so tsc is used merely for type-checking.
Setting noEmit to true makes sure that other tools which use tsc (like
editor plugins) do not make tsc compile code.

outDir is removed as it's not needed if we do not use tsc to compile code.