Add semver equivalents of teleport.Version and api.Version (#38038) (#38153)
The teleport.MinClientVersion was also updated to be derived from
the new semver variant as well as adding a semver.Version variant.
Additionally instances of semver.New(teleport.Version) or
semver.New(api.Version) were replaced to use the newly introduced
variables. It was also noticed that auth.MinSupportedModeratedSessionsVersion
was no longer being used so it was removed.