Prevent flaky behavior in TestTerminalSizeRequest (#38227)
The test flakiness was derived from the asynchronous nature of
processing window change requests. The client was sending a window
change request and then immediately sending another request to retreive
the window size and expected the server to have already processed
the window change request. Now the test is sending a window change
request, waiting for the server to finish processing the request,
and then sending a request to get the terminal size.

Fixes #36062