ci: Make relcli docker image work in GitHub Actions (#3451)
* ci: Make relcli docker image work in GitHub Actions

Use the relcli OCI image via `docker run` instead of as a GitHub Actions
container job. Running as a container job requires the `tail` program be
in the image as GHA runs `tail -f /dev/null` for the duration of the
job. The relcli image does not have `tail` so this fails.

This allows us to consolidate the two jobs as previously we could not
use the `aws` command to get the ECT login creds. We no longer need the
creds directly as within the single job we can use the amazon-ecr-login
action to autheticate to ECR.

* fixup! ci: Make relcli docker image work in GitHub Actions

* fixup! ci: Make relcli docker image work in GitHub Actions

Source-commit: gravitational/teleport.e@56535aaa09872ec27d965562acabb3e04e01a5f0
Source-pull-requests: gravitational/teleport.e#3451