ci: Add top-level workflow for publishing a tag build (#3452)
* ci: Make publish (promote) workflows callable (workflow_call)

Add a `workflow_call` trigger to the following workflows:
* publish-helm
* publish-to-houston
* publish-os-packages
* promote-teleport-kube-agent-updater-oci
* promote-teleport-spacelift-runner-oci
* promote-teleport-oci-distroless
* promote-teleport-hardened-amis
* publish-release-server

The `workflow_call` trigger duplicates the inputs of the existing
`workflow_dispatch` trigger. In some rare cases the type of the
parameter has been changed to `string` where the type on the dispatch
parameter is not supported by call events. The descriptions have been
removed from the `workflow_dispatch` event to simplify maintenance.

It is expected that the `workflow_dispatch` event will be removed when
Drone no longer needs to call it.

Clean up the input parameters of both by unquoting strings, using a
single line for description where it fits, ensuring the description ends
with a period, ordering the input keys as description, type,
default/required, remove `required` where it has a default, removing
keys with values that are the default.

* ci: Update job names for clearer display

Update the job name for oci-image-promote to show the pending OCI
repository that is being published instead of the tag version. When
publishing, everything artifact is the same version, so showing that for
each artifact being published does not differentiate what is being
published. The pending repository does.

Add a name to the deploy-teleport-artifacts job. Without it, GitHub
actions uses the job name and the matrix vars, which is all a bit long
and the pertinent stuff gets truncated. This should make each sub-job
clearer in the list.

* ci: Add top-level workflow for publishing a tag build

Add a top-level `tag-publish` workflow that publishes the pending
artifacts from a tag build (`tag-build`). It takes the place of the
Drone pipelines that trigger off the `promote` event.

Currently this workflow is manually dispatched (`workflow_dispatch`)
trigger, either manually from the GitHub web UI, manually via the `gh`
command or soon a target in the `teleport` `Makefile`.

A sample `gh` command to trigger the workflow is:

    gh workflow run -R gravitational/teleport.e tag-publish.yaml \
      -f oss-teleport-ref=v15.0.1

To trigger a dev build from master with a custom version:

    gh workflow run -R gravitational/teleport.e tag-publish.yaml \
      -f oss-teleport-ref=master -f teleport-version=16.0.0-dev.camh.1

To trigger publishing a private release:

    gh workflow run -R gravitational/teleport.e tag-publish.yaml \
      -f oss-teleport-repo=gravitational/teleport-private \
      -f oss-teleport-ref=v15.0.1

Source-commit: gravitational/teleport.e@9c3bb7c624a3cbaf40291087764e249d8e25dd8e
Source-pull-requests: gravitational/teleport.e#3452