[v15] AWS OIDC Service: bootstrap and ListDatabases (#37785)
* AWS OIDC Service: bootstrap and ListDatabases stub

This PR adds a new gRPC Service to handle the AWS OIDC Integration
actions.

All the HTTP exposed actions will use this new service.
Eg, list databases, list ec2 instances, and so on

This PR adds the new service and a placeholder for the first action to
me refactored: ListDatabases.
This action is not used/implemented yet.

* AWSOIDC: ListDatabases service (#37853)

* AWSOIDC: ListDatabases service

This PR implements the ListDatabases service.
The Web API will now use this service instead of calling the
GenerateAWSOIDC Token.

* log and skip instead of failing when convertion fails