Fix hostname resolution for connections in leaf clusters
The changes to hostname resolution introduced in #35773 were always
using the root cache to lookup a matching node, which is never
going to find a match if the target is in the leaf cluster. This
is now resolved by using the `services.NodeWatcher` of the cluster
that the target node is a member of.