ci: Add publish-operator-oci to tag-publish workflow (#3467)
* ci: Fix logging into public ECR for publish-operator-oci

Fix the `registry-type` and region when logging into the destination
registry. The previous attempt was trying to be clever to allow
publishing to a private repo, for testing. But this all works properly
in the staging environment with a public repo and that is better for
proper testing anyway, so just hard-code the public registry. Also add
the region which I thought was automatic when the registry type is
public, but apparently not.

(Let's just ignore the incorrect order of args to `contains` as well and
pretend that never happened).

This has all been tested - I fixed this and tested it a week and a half
ago but inadvertently lost it with a git muck-up and unknowingly merged
an old version.

* ci: Make publish-operator-oci callable

Add a `workflow_call` trigger to the `publish-operator-oci` workflow so
it can be called from the `tag-publish` workflow. This duplicates the
`workflow_dispatch` input parameters, but that trigger is left so we can
manually call the single workflow for testing.

* ci: Add publish-operator-oci to tag-publish

Call the `publish-operator-oci` from the `tag-publish`. This was
accidentally left out in the commit that added all the other jobs.

* ci: Use production environment for publishing OS packages

Select the `environment` for the `publish-os-packages` workflow based on
whether the version being published has a pre-release component in the
semver. Pre-release versions go to `development`, full releases go to
`production`. This was meant to be done in a previous commit but was
missed.

Note: `production` and `development` are a pseudo-environments here -
the called workflow maps that to an actual GitHub Actions environment.

Source-commit: gravitational/teleport.e@fa415fd31b30f704be63095d30b438422fa6b045
Source-pull-requests: gravitational/teleport.e#3467