ci: Add publish-operator-oci to release dependencies (#3487)
Add the `publish-operator-oci` job as a dependency of the
`publish-release-server` job. This was missed when this job was added.

While we're here, add a little bit of debugging and a GitHub step
summary for what images were published and what were skipped. This makes
it a little clearer what's going on.

Source-commit: gravitational/teleport.e@84b543b99a7c9500c39435abc257448da416fe84
Source-pull-requests: gravitational/teleport.e#3487