Add cluster config service proto (#38208)
The new RPC service will replace the legacy RPCs in proto.AuthService
so that cluster config resources can be protected by optimistic locking.
The current API for ClusterNetworkingConfig, SessionRecordingConfig,
ClusterAuditConfig and AuthPreferences is limited to Get/Set which
makes it impossible to support optimistic locking.

These are all singleton resources whose creation is managed by
Auth and deletion is not allowed. Due to these constraints there
are no Create RPCs exposed to users, though there is an Upsert
for use by tctl create -f, and no Delete RPCs, instead there is
a Reset which resturns the configuration to the default value.