ci: Add dummy workflow for building buildboxes (#3501)
Add a dummy workflow so a workflow to build the buildboxes in teleport.e
can be developed. The workflow must exist on master in some form for a
`workflow_dispatch` trigger to be called, on any branch.

It is not clear if building the buildboxes will be ultimately done from
within a teleport.e workflow, but right now only this repository can
access the self-hosted arm64 runners needed to build the arm64
buildboxes. While there is a private beta of GitHub-hosted arm64
runners, it is not clear yet how ready for use they are, so this stands
in as an alternative.

Source-commit: gravitational/teleport.e@415d2f9151a9cbcb8ccb3167264b12d1e2ede61b
Source-pull-requests: gravitational/teleport.e#3501