Remove `inetutils-telnet` from legacy Ubuntu OCI due to CVE-2021-40491 (#38506)
Ubuntu only fixed CVE-2021-40491 in their paid ESM offering (as `inetutils` is
part of `universe`, not `main`), and it's causing security scanners to complain.

While the recommended option is to move to our distroless OCIs which do
not have this package installed at all (and are updated daily), an easy fix is
just to remove `inetutils-telnet` from the image (as `netcat` is already included).

References:
* https://ubuntu.com/security/CVE-2021-40491
* https://ubuntu.com/security/notices/USN-5177-1
* https://ubuntu.com/security/esm

changelog: Remove `telnet` from legacy Ubuntu OCI due to CVE-2021-40491. Use `nc` instead.