Add the cluster configuration RPC service
The new clusterconfigv1.Service implements the gRPC service defined
in teleport.clusterconfig.v1.ClusterConfigService. The initial
implementation is restricted to supporting cluster auth preferences.
Support for remaining cluster configuration resources will come
in a series of future PRs. Once all of the resources are covered
the service will be added to the Auth server and consumers of the
old API will be migrated.