Perform server version validation before using clients (#38457)
Agents performed the following steps when creating an auth client:

1) Create an auth gRPC client
2) Call proto.AuthService/GetDomainName to force gRPC to dial the
   auth server in order to validate the connection since gRPC only
   dials on the first RPC
3) Call proto.AuthService/GetCertAuthority to get the HostCA
4) Call proto.AuthService/Ping to get the server version to determine
   if the connection should be aborted unless --skip-version-check
   was provided

On the surface this seems fine, however, the skip version check
did not happen until after two RPCs to auth had already been issued.
If either of those RPCs has changed the agent could perpetually fail
to connect to the cluster even though it should have never tried
due to version differences.

The order of the steps was changed to the following such that the
version check happens first.

1) Create an auth gRPC client
2) Call proto.AuthService/Ping to force gRPC to dial AND to get the
   server version information
3) Run the server version check and abort unless --skip-version-check
   was provided
4) Call proto.AuthService/GetCertAuthority to get the HostCA

An added benefit to this is agents now require one less RPC when
joining.