ci: Switch tag build workflows to artifact@v4 GHA action (#3510)
Switch the mac and linux tag build workflows to use
`actions/upload-artifact@v4` and `actions/download-artifact@v4`, from
v3. This is primarily as we need to rename the artifacts so we can call
mulitple workflows from a single top-level workflow - artifact names are
scoped to the top workflow, so these workflows which were all using the
same `release` artifact name would conflict.

Since this needs to be fixed, this is a good opportunity to upgrade to
v4 of the artifact actions which is meant to be a lot faster (it was so
slow before it would be hard to be slower - like 5m to upload artifacts
which take 2s to upload to S3).

This change changes the input parameter `save-artifacts` to a string
parameter specifying the artifact name, where an empty name means not to
save the artifacts.

We also specify a compression level of 0 as the artifacts are already
compressed, so there's no point wasting CPU trying to compress them
again.

Source-commit: gravitational/teleport.e@21a5a66fa79c288311ba22441dcd962170bfe0fb
Source-pull-requests: gravitational/teleport.e#3510