ci: Fix RPM build with release-amd64{,-fips} targets (#3500)
* build: Generate centos7 archive when building linux-amd64

When building a linux-amd64 archive, make a copy of it with the
`centos7` tag as the linux-amd64 build works on centos7. We stopped
doing a centos7-specific build a while ago, but we still have the
archive on our releases page.

This helps unify the `release-amd64` and `release-amd64-centos7`
targets, which currently do the same thing except for the GitHub Actions
workflow that creates the centos7 archive only for the latter target.
This will allow us to get rid of that target as that latter target will
no longer be called when Drone is removed.

Also add a `-p` when making RELEASE_DIR as sometimes the parent
directory has not yet been created.

* ci: Stop generating centos7 tarball in build-linux workflow

Remove the custom logic in the `build-linux` workflow for creating a
copy of the linux binary archive for centos7, as this has moved to the
`Makefile`.

Simplify the remaining logic to copy anything in `e/build/artifacts` to
the top-level `build/artifacts` which will allow changes to the Makefile
to just put everything in the top-level `build/artifacts`
(`$RELEASE_DIR`), after which we can remove this logic from the workflow
altogether.

Source-commit: gravitational/teleport.e@9e533867bccd234c99c3cb7df45c143d235877a7
Source-pull-requests: gravitational/teleport.e#3500