ci: Add top-level workflow for building a tagged release (#3511)
* ci: Make release workflows callable (workflow_call)

Add a `workflow_call` trigger to the following workflows:
* release-linux
* release-mac
* release-windows
* release-teleport-kube-agent-updater-oci
* release-teleport-spacelift-runner-oci
* release-teleport-oci-distroless
* release-teleport-hardened-ami

The `workflow_call` trigger duplicates the inputs of the existing
`workflow_dispatch` trigger. In some rare cases the type of the
parameter has been changed to `string` where the type on the dispatch
parameter is not supported by call events. The descriptions have been
removed from the `workflow_dispatch` event to simplify maintenance.

It is expected that the `workflow_dispatch` event will be removed when
Drone no longer needs to call it.

Clean up the input parameters of both by unquoting strings, using a
single line for description where it fits, ensuring the description ends
with a period, ordering the input keys as description, type,
default/required, remove `required` where it has a default, removing
keys with values that are the default.

* ci: Turn off fail-fast for release-mac workflow

Turn off the `fail-fast` setting when running the per-architecture
matrix builds so that a transient failure in one does not cancel the
other. Often failures are for transient reasons, so we can just re-run
the failed jobs to continue the workflow, and that will go faster if we
don't have to re-run the cancelled jobs. For the case of a genuine build
failure due to bad code or similar, it is likely the other job(s) will
hit the same condition and fail soon anyway.

* ci: Add top-level workflow for building a tagged release

Add a top-level `tag-build` workflow that builds a tagged release,
pushing all built artifacts to pending repositories (S3, ECR, AMI, etc)
ready to be published by the `tag-publish` workflow. It takes the place
of the Drone pipelines that trigger off the `tag` event.

Currently this workflow is manually dispatched (`workflow_dispatch`
trigger), either manually from the GitHub web UI, manually via the `gh`
command or soon a target in the `teleport` `Makefile`.

Some sample `gh` commands to trigger the workflow are:

    gh workflow run -R gravitational/teleport.e tag-build.yaml \
      -f oss-teleport-ref=v15.0.1

To trigger a dev build from master with a custom version:

    gh workflow run -R gravitational/teleport.e tag-build.yaml \
      -f oss-teleport-ref=master -f teleport-version=16.0.0-dev.camh.1

To trigger building a private release:

    gh workflow run -R gravitational/teleport.e tag-build.yaml \
      -f oss-teleport-repo=gravitational/teleport-private \
      -f oss-teleport-ref=v15.0.1

Source-commit: gravitational/teleport.e@38c600b15594c174dc551cffdf969ffc35c04cc2
Source-pull-requests: gravitational/teleport.e#3511