Introduce member_count field into access lists.
A count of the members in an access listhas been introduced into access
This field should not be populated by the backend, but must be populated by
the gRPC call to determine whether the user should have access to this data.
Additionally, this field is excluded from the JSON/YAML deserialization, so
the field won't propagate to the backend. It will only be used/seen by API
callers.