Include an optional screen size in the WindowsDesktop spec (#38307)
By default this should remain unspecified, which preserves the
behavior we have today (Teleport uses the size of the user's
browser window).

These new fields allow cluster administrators to configure desktops
in such a way that they always use a particular screen size, ignoring
the size passed from the browser client. This is necessary to
accommodate some applications which require a very specific screen resolution in
order to render correctly.