Fix sticky windows key (#38198)
* Adds the extended keycode where appropriate and updates to the latest IronRDP master commit hash.

* cleaning up disconnect result

* Pipes disconnect messages through to the client.

This required a change to the state management for DesktopSession. State
is still quite complex to get right here, but this is a step in the right
direction.

* Updates desktopPlaybackHandle to the new authenticated websocket paradigm as it was forgotten in https://github.com/gravitational/teleport/pull/37520

* prettier formatting

* adding clarifying comments

* Further refactoring to make DesktopSession's state machine more comprehensible

* Remove unnecessary items from dependency lists

* Adds the KeyboardHandler class
that manages the keyboard events, including new functionality to handle
withholding the Windows/Alt keys to prevent such keys from "sticking"
when the user switches between windows.

This technique works to stop the sticking, though there is still an edge
case where the user cmd/alt-tabs super quickly, in a manner that the browser
sees both the down and up events for the Windows key, despite the user's
intention being to change windows. In the case of the Windows key, this
causes the start menu to pop open, and the user must press the Windows key
again to close it. This is not a showstopper, but it is a bit annoying.
A similar problem would occur with the Alt key, where such behavior would
cause the current window (on the Windows machine) to focus on the menu bar.

The next commit will be to add a timeout mechanism for the relevant keyup events
in order to mitigate this edge case.

* Adds a delayUp cache for delaying the up event
of the sticky windows key. This is to prevent the edge case
where the user presses cmd/alt-tab really quickly in a manner
that causes the down + up events to be registered in spite of
the user's intention to be to change out of the window.

* Refactoring KeyboardHandler

* Convert caches into Maps

* Moves the rest of keyboard functionality into KeyboardHandler, rearranges methods, adds an onUnmount for clearing all the intervals on unmount

* Moves KeyboardHandler to its own file

* simplifies withholding logic by including a Withholder class with a single array that manages all the withheld keystrokes

* Makes useEffect symmetrical

* Rename onUnmount to dispose

* sp

* Adds unit tests for Withholder

* Splitting Withholder out into its own file

* nits

* CR

* make fix-license