Add admin user in the `teleport-cluster` guide
Closes #10583

The local user we show how to create in the `teleport-cluster` guide
only has privileges to access Kubernetes resources, but does not have
administrative access to the cluster (e.g., to manage dynamic
resources).

This change takes the approach of the Linux server guide (the index page
of the docs) and shows how to add a user with the `editor,access` roles
in addition to a role that can access Kubernetes clusters.